+++ nautilus_core/serialization/src/arrow/depth.rs
@@ -386,7 +386,7 @@ mod tests {
     use arrow::datatypes::{DataType, Field};
     use nautilus_model::{
         data::stubs::stub_depth10,
-        types::{fixed::SCALAR, price::PriceRaw, quantity::QuantityRaw},
+        types::{fixed::FIXED_SCALAR, price::PriceRaw, quantity::QuantityRaw},
     };
     use pretty_assertions::assert_eq;
     use rstest::rstest;
@@ -492,7 +492,7 @@ mod tests {
             assert_eq!(bid_price.len(), 1);
             assert_eq!(
                 get_raw_price(bid_price.value(0)),
-                (expected_bid_prices[i] * SCALAR) as PriceRaw
+                (expected_bid_prices[i] * FIXED_SCALAR) as PriceRaw
             );
             assert_eq!(
                 Price::from_raw(get_raw_price(bid_price.value(0)), price_precision).as_f64(),
@@ -518,7 +518,7 @@ mod tests {
             assert_eq!(ask_price.len(), 1);
             assert_eq!(
                 get_raw_price(ask_price.value(0)),
-                (expected_ask_prices[i] * SCALAR) as PriceRaw
+                (expected_ask_prices[i] * FIXED_SCALAR) as PriceRaw
             );
             assert_eq!(
                 Price::from_raw(get_raw_price(ask_price.value(0)), price_precision).as_f64(),
@@ -540,7 +540,7 @@ mod tests {
             assert_eq!(bid_size.len(), 1);
             assert_eq!(
                 get_raw_quantity(bid_size.value(0)),
-                ((100.0 * SCALAR * (i + 1) as f64) as QuantityRaw)
+                ((100.0 * FIXED_SCALAR * (i + 1) as f64) as QuantityRaw)
             );
         }
 
@@ -558,7 +558,7 @@ mod tests {
             assert_eq!(ask_size.len(), 1);
             assert_eq!(
                 get_raw_quantity(ask_size.value(0)),
-                ((100.0 * SCALAR * ((i + 1) as f64)) as QuantityRaw)
+                ((100.0 * FIXED_SCALAR * ((i + 1) as f64)) as QuantityRaw)
             );
         }